+++ bootstrap.sh
@@ -14,7 +14,7 @@ if [ -x "$(which python)" ]; then
     ./py2_zodb3/bin/pip install https://github.com/bluedynamics/odict/archive/master.zip
     ./py2_zodb3/bin/pip install https://github.com/bluedynamics/plumber/archive/master.zip
     ./py2_zodb3/bin/pip install https://github.com/bluedynamics/node/archive/master.zip
-    ./py2_zodb3/bin/python setup.py develop
+    ./py2_zodb3/bin/pip install -e .
 
     rm -r py2_zodb5
 
@@ -29,7 +29,7 @@ if [ -x "$(which python)" ]; then
     ./py2_zodb5/bin/pip install https://github.com/bluedynamics/odict/archive/master.zip
     ./py2_zodb5/bin/pip install https://github.com/bluedynamics/plumber/archive/master.zip
     ./py2_zodb5/bin/pip install https://github.com/bluedynamics/node/archive/master.zip
-    ./py2_zodb5/bin/python setup.py develop
+    ./py2_zodb5/bin/pip install -e .
 fi
 if [ -x "$(which python3)" ]; then
     rm -r py3_zodb5
@@ -45,5 +45,5 @@ if [ -x "$(which python3)" ]; then
     ./py3_zodb5/bin/pip install https://github.com/bluedynamics/odict/archive/master.zip
     ./py3_zodb5/bin/pip install https://github.com/bluedynamics/plumber/archive/master.zip
     ./py3_zodb5/bin/pip install https://github.com/bluedynamics/node/archive/master.zip
-    ./py3_zodb5/bin/python setup.py develop
+    ./py3_zodb5/bin/pip install -e .
 fi